Add Support for buttons and clean code a bit
